refactor Database and DatabaseHolder into /database (#228)
* refactor Database and DatabaseHolder into /database

* refactor FirestoreDatabase.kt into /database

* refactor SnapshotUtility.kt into /database

* refactor SnapshotUtility.kt into /database
