+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockJoinGameActivities.kt
@@ -3,7 +3,7 @@ package com.github.fribourgsdp.radio.mockimplementations
 import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.DialogFragment
-import com.github.fribourgsdp.radio.Database
+import com.github.fribourgsdp.radio.database.Database
 import com.github.fribourgsdp.radio.game.prep.JoinGameActivity
 import com.github.fribourgsdp.radio.data.User
 import org.mockito.Mockito.*
