+++ app/src/main/java/com/github/fribourgsdp/radio/data/Playlist.kt
@@ -1,7 +1,7 @@
 package com.github.fribourgsdp.radio.data
 
-import com.github.fribourgsdp.radio.Database
-import com.github.fribourgsdp.radio.FirestoreDatabase
+import com.github.fribourgsdp.radio.database.Database
+import com.github.fribourgsdp.radio.database.FirestoreDatabase
 import com.github.fribourgsdp.radio.util.SetUtility
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
