+++ app/src/main/java/com/github/fribourgsdp/radio/data/User.kt
@@ -3,7 +3,7 @@ package com.github.fribourgsdp.radio.data
 
 import android.content.Context
 import com.github.fribourgsdp.radio.persistence.*
-import com.github.fribourgsdp.radio.FirestoreDatabase
+import com.github.fribourgsdp.radio.database.FirestoreDatabase
 import com.github.fribourgsdp.radio.util.SetUtility
 
 import com.google.android.gms.tasks.Task
