+++ app/src/main/java/com/github/fribourgsdp/radio/database/DatabaseHolder.kt
@@ -1,0 +1,7 @@
+package com.github.fribourgsdp.radio.database
+
+interface DatabaseHolder {
+    fun initializeDatabase() : Database {
+        return FirestoreDatabase()
+    }
+}
