+++ app/src/main/java/com/github/fribourgsdp/radio/deprecated/FireBaseTestActivity.kt
@@ -2,7 +2,7 @@ package com.github.fribourgsdp.radio.deprecated
 
 import android.content.Intent
 import android.os.Bundle
-import com.github.fribourgsdp.radio.FirestoreDatabase
+import com.github.fribourgsdp.radio.database.FirestoreDatabase
 import com.github.fribourgsdp.radio.MainActivity
 import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.config.MyAppCompatActivity
