+++ app/src/main/java/com/github/fribourgsdp/radio/game/prep/JoinGameActivity.kt
@@ -13,8 +13,8 @@ import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.DialogFragment
-import com.github.fribourgsdp.radio.Database
-import com.github.fribourgsdp.radio.FirestoreDatabase
+import com.github.fribourgsdp.radio.database.Database
+import com.github.fribourgsdp.radio.database.FirestoreDatabase
 import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.User
